Remove signed update system entirely
The point of assymetric crypto was for the CM to be able to prove
authorship to the arbiter by signing updates. This commit makes the CM
instead use the key system that everything else uses. The key is passed
to the arbiter as an environment variable just as it is for other
containers. Since we use HTTPS for all connections to the arbiter, the
keys are protected as they are passed with requests. Resolves #11.