Enable communication with the arbiter and other containers
The container manager uses a proxy to map requests to docker ephemeral  ports on the localhost. This was broken due to the move to https.  The containers IP did not match the HTTPS certs host.  This approach would also not work with MacOS or Windows. 
 
To fix this the container managers DNS lookups are being intercepted. If a lookup fails to resolve they are mapped to 127.0.0.1. This fixes the HTTPS host mismatch