Fixed HTTPS problems and moved to databoxsystems repo
Moving to localhost for registryUrl_dev and storeUrl_dev fixed some problems but created a monster. The when the CM tries to access  localhost:5000 it resolves to the container, not the host ;-(

This means in some cases the HTTPS cert needs to be valid for localhost and in others, it needs to be valid for the container name. I have added localhost as an Alternative Name (SAN) to all certs and fiddled with when we pass localhost and when we pass the container name in the code. 

This should hopefully be less of a problem in future when the CM UI is hosted off the databox.