Pass hypercat item-metadata to apps at install time
The app gets passed information about the data sources they have been assigned at install time. This data is passed through an environment variable. 

Until now the data passed to the app was a custom JSON object encoded as a string. From now on the data passed will be the hypercat item relating to that data source. The hypercat item is extracted from the datastores /cat endpoint by the container manger.