databox with core-network (#155)
* start databox with bridge

* use dnsrr instead of vip

* Pin core-container-manager to sevenEng/core-container-manager

This will make the test run using your modified container-manager

* Add sevenEng/databox-bridge

so that the image is built from src for the tests

* using with-bridge branch of CM

* Build the databox-bridge

I really should fix this so it uses the core-components file. Instead of hardcoding in here. 

Note: It is hardcode because the image names don't match the repo names.

* databox-bridge is used in start script as image name

* use docker-compose to start/stop bridge

* comply to travis supported compose file version, up to 3.2

* put export-service on a separate external-accessible net

* rename some bridge to core-network

* get rid of all the 'bridge'

* make system-net not being internal

* add aliase to sync with altname in root ca